protect update and delete route if logged in user is not owner

+++ app/Http/Controllers/ListingController.php
@@ -101,6 +101,11 @@ class ListingController extends Controller
      */
     public function update(Request $request, Listing $listing)
     {
+        // make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             "title" => "required",
             "company" => "required",
@@ -128,6 +133,11 @@ class ListingController extends Controller
      */
     public function destroy(Listing $listing)
     {
+        // make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted sucessfully');
     }
